+++ tsdb/aggregate.go
@@ -342,11 +342,12 @@ func (e *AggregateExecutor) processFill(results [][]interface{}) [][]interface{}
 		return newResults
 	}
 
+	isCount := e.stmt.HasSimpleCount()
 	// They're either filling with previous values or a specific number
 	for i, vals := range results {
 		// start at 1 because the first value is always time
 		for j := 1; j < len(vals); j++ {
-			if vals[j] == nil {
+			if vals[j] == nil || (isCount && isZero(vals[j])) {
 				switch e.stmt.Fill {
 				case influxql.PreviousFill:
 					if i != 0 {
@@ -359,6 +360,18 @@ func (e *AggregateExecutor) processFill(results [][]interface{}) [][]interface{}
 		}
 	}
 	return results
+}
+
+// Returns true if the given interface is a zero valued int64 or float64.
+func isZero(i interface{}) bool {
+	switch v := i.(type) {
+	case int64:
+		return v == 0
+	case float64:
+		return v == 0
+	default:
+		return false
+	}
 }
 
 // processDerivative returns the derivatives of the results
@@ -694,7 +707,7 @@ func (m *AggregateMapper) initializeMapFunctions() error {
 		}
 		m.mapFuncs[i] = mfn
 
-		// Check for calls like `derivative(lmean(value), 1d)`
+		// Check for calls like `derivative(mean(value), 1d)`
 		var nested *influxql.Call = c
 		if fn, ok := c.Args[0].(*influxql.Call); ok {
 			nested = fn